requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Session==0.8.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.1
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
supabase==2.9.1
twilio==9.3.7
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
# Cashfree Payment Gateway
cashfree-pg==4.0.4

# Background Scheduler for Monthly Rollover
APScheduler==3.10.4

# Celery and Redis for async processing
celery==5.5.3
redis==6.2.0
anthropic==0.57.1
supervisor==4.2.5
psutil==5.9.6
kombu==5.5.4
billiard==4.2.1
vine==5.1.0

# Real-time progress tracking
Flask-SocketIO==5.4.1

# PostgreSQL adapter for direct database connections
psycopg2-binary==2.9.9

# Firebase Admin SDK for OTP authentication
firebase-admin==6.5.0
